+++ app/controllers/api/dictionaries_controller.rb
@@ -30,9 +30,13 @@ module Api
       @dictionaries = if query.empty?
                         Dictionary.all.order(:condition)
                       else
-                        Dictionary.all.select do |dictionary|
-                          query[:content].include?(dictionary.phrase)
-                        end.sort_by(&:condition)
+                        dictionaries = Dictionary.where(query.except(:content))
+                        if query.key?(:content)
+                          dictionaries = dictionaries.select do |dictionary|
+                            query[:content].include?(dictionary.phrase)
+                          end
+                        end
+                        dictionaries.sort_by(&:condition)
                       end
       render status: :ok, template: 'dictionaries/dictionaries'
     end
@@ -44,7 +48,7 @@ module Api
     end
 
     def index_params
-      %i[content]
+      %i[condition content phrase]
     end
   end
 end