+++ app/models/dictionary.rb
@@ -1,10 +1,10 @@
-class Dictionary < ActiveRecord::Base
+class Dictionary < ApplicationRecord
   has_many :category_dictionaries, dependent: :destroy
   has_many :categories, through: :category_dictionaries
 
-  validates :phrase,
-            presence: {message: 'invalid'}
-
+  validates :phrase, :condition,
+            presence: {message: 'absent'}
   validates :condition,
-            inclusion: {in: %w[equal include], message: 'invalid'}
+            inclusion: {in: %w[equal include], message: 'invalid'},
+            allow_nil: true
 end