+++ spec/models/payment_spec.rb
@@ -35,7 +35,7 @@ describe Payment, type: :model do
         ],
       ].each do |interval, settlement|
         context "#{interval}を指定する場合" do
-          include_context '事前準備: 収支情報を登録する', [income, expense]
+          include_context '収支情報を登録する', [income, expense]
           before(:all) { @settlement = Payment.settle(interval) }
 
           it '計算結果が正しいこと' do
@@ -55,52 +55,27 @@ describe Payment, type: :model do
   end
 
   describe '#validates' do
-    valid_params = {
-      payment_type: 'income',
-      date: '1000-01-01',
-      content: 'モジュールテスト用データ',
-      price: 1000,
-    }
-
-    shared_context 'Paymentオブジェクトを検証する' do |params|
-      before(:all) do
-        @payment = Payment.new(params.except(:category))
-        @payment.validate
-      end
-    end
-
-    shared_examples '検証結果が正しいこと' do |result|
-      it_is_asserted_by { @payment.errors.empty? == result }
-    end
-
     describe '正常系' do
-      %w[1000-01-01 1000/01/01 01-01-1000 01/01/1000 10000101].each do |date|
-        context "date=#{date}の場合" do
-          include_context 'Paymentオブジェクトを検証する', valid_params.merge(date: date)
-          it_behaves_like '検証結果が正しいこと', true
-        end
-      end
+      valid_attribute = {
+        payment_type: %w[income expense],
+        date: %w[1000-01-01 1000/01/01 01-01-1000 01/01/1000 10000101],
+        content: 'モジュールテスト用データ',
+        price: 0,
+      }
+
+      it_behaves_like '正常な値を指定した場合のテスト', valid_attribute
     end
 
     describe '異常系' do
-      invalid_params = {
-        payment_type: 'invalid_type',
-        date: ['invalid_date', '1000-13-01', '1000-01-00', '1000-13-00'],
+      invalid_attribute = {
+        payment_type: 'invalid',
+        date: %w[invalid 1000-13-01 1000-01-00 1000-13-00],
         price: [-1],
       }
+      absent_keys = %i[payment_type content price]
 
-      CommonHelper.generate_test_case(invalid_params).each do |params|
-        context "#{invalid_params.keys.join(',')}が不正な場合" do
-          include_context 'Paymentオブジェクトを検証する', valid_params.merge(params)
-
-          it_behaves_like '検証結果が正しいこと', false
-
-          it 'エラーメッセージが正しいこと' do
-            error_messages = params.map {|key, _| [key, ['invalid']] }.to_h
-            is_asserted_by { @payment.errors.messages == error_messages }
-          end
-        end
-      end
+      it_behaves_like '必須パラメーターがない場合のテスト', absent_keys
+      it_behaves_like '不正な値を指定した場合のテスト', invalid_attribute
     end
   end
 end